Move quote ingestion into API app
