+++ apps/atomy-q/API/IMPLEMENTATION_SUMMARY.md
@@ -379,7 +379,7 @@ Quote intake persistence is now tenant-scoped for `upload`, `index`, and `show`:
 
 **Quote intake + normalization (live):** `QuoteSubmissionController` now persists live upload/show/reparse data, returns quote-submission state with processing metadata, and requeues reparses against the tenant-scoped row. `NormalizationSourceLine` and `NormalizationConflict` models match the shipped migrations (`rfq_line_item_id`, `normalization_source_line_id`, etc.). `QuoteSubmissionReadinessService` centralizes blocking rules: each RFQ line must be covered by a mapped source line, mapped lines require `source_unit_price`, and open conflicts (`resolution` null) block readiness. `NormalizationController` persists mapping/bulk-mapping/override/revert/conflict-resolution, returns `404` when the RFQ or rows are not tenant-visible, and recalculates `quote_submissions.status` to `needs_review` or `ready` after mutations. `GET /normalization/{rfqId}/conflicts` lists conflicts for the RFQ and includes `meta.has_blocking_issues` / `meta.blocking_issue_count` aggregated across that RFQ’s submissions. Feature coverage: `QuoteIngestionPipelineTest`, `NormalizationReviewWorkflowTest`.
 
-**Quote ingestion + normalization hardening (April 2026):** `QuoteIngestionOrchestrator` now delegates into `QuotationIntelligenceCoordinator` with tenant-context propagation, RFQ-aware decision-trail writes, and delta reparse preservation for manual overrides. The API app binds tenant-scoped bridge adapters for document lookup, tenant lookup, RFQ procurement lookup, decision-trail persistence, deterministic content extraction, deterministic semantic mapping, in-memory UoM lookup, and static exchange-rate lookup. `NormalizationController` now serializes `ai_confidence`, `taxonomy_code`, and `mapping_version` so the UI can render the intelligence payload without extra joins. Feature coverage now includes `QuoteIngestionIntelligenceTest` and `QuoteIngestionPipelineTest`, both passing on isolated SQLite verification runs.
+**Quote ingestion + normalization hardening (April 2026):** `App\Services\QuoteIntake\QuoteIngestionOrchestrator` now delegates into `QuotationIntelligenceCoordinator` with tenant-context propagation, RFQ-aware decision-trail writes, and delta reparse preservation for manual overrides. The API app binds tenant-scoped bridge adapters for document lookup, tenant lookup, RFQ procurement lookup, decision-trail persistence, deterministic content extraction, deterministic semantic mapping, in-memory UoM lookup, and static exchange-rate lookup. `NormalizationController` now serializes `ai_confidence`, `taxonomy_code`, and `mapping_version` so the UI can render the intelligence payload without extra joins. Feature coverage now includes `QuoteIngestionIntelligenceTest` and `QuoteIngestionPipelineTest`, both passing on isolated SQLite verification runs.
 
 **PetrochemicalTenantSeeder update (2026-04-27):** Updated seeder to mimic quotations through multiple processing stages (normalization, extraction, etc.) with new DB fields including `error_code`, `error_message`, `processing_started_at`, `processing_completed_at`, `parsed_at`, `retry_count` for submissions, and `ai_confidence`, `taxonomy_code`, `mapping_version` for normalization lines.
 
@@ -387,7 +387,7 @@ Quote intake persistence is now tenant-scoped for `upload`, `index`, and `show`:
 
 **Quote intelligence review follow-up (2026-04-16):** The deterministic processor now tenant-scopes the parent RFQ eager load in addition to tenant-scoping the quote submission and RFQ line items, the deterministic semantic mapper is declared `readonly` and documents its intentionally tenant-agnostic validation path, and the operator docs now spell out the `QUOTE_INTELLIGENCE_*` env contract plus the logged, sanitized failure behavior for dormant `llm` mode.
 
-**Quote ingestion retry failure persistence (2026-04-16):** `ProcessQuoteSubmissionJob::failed()` now logs the full retry-exhaustion exception server-side, then persists `status = failed`, `error_code = MAX_RETRIES_EXCEEDED`, and the sanitized generic `QuoteIngestionOrchestrator::GENERIC_FAILURE_MESSAGE` on the submission row so API responses do not leak backend details.
+**Quote ingestion retry failure persistence (2026-04-16):** `ProcessQuoteSubmissionJob::failed()` now logs the full retry-exhaustion exception server-side, then persists `status = failed`, `error_code = MAX_RETRIES_EXCEEDED`, and the sanitized generic `App\Services\QuoteIntake\QuoteIngestionOrchestrator::GENERIC_FAILURE_MESSAGE` on the submission row so API responses do not leak backend details.
 
 **Comparison snapshot, preview, matrix, award, and approval (live):** `ComparisonRun` / `Approval` / `DecisionTrailEntry` models align with migrations (`response_payload`, `requested_by`, hash-chain columns). `ComparisonRunController` now binds `BatchQuoteComparisonCoordinatorInterface`, `ComparisonReadinessValidatorInterface`, `QuoteComparisonMatrixServiceInterface`, and `ApprovalGateServiceInterface` so preview/matrix/readiness use the live `Nexus\QuotationIntelligence` comparison boundary. `POST /comparison-runs/preview` persists a real preview `ComparisonRun` row, stores live `matrix_payload` / `scoring_payload` / `approval_payload` / `readiness_payload`, and returns the persisted run id instead of a synthetic preview token. `GET /comparison-runs/{id}/matrix` and `GET /comparison-runs/{id}/readiness` return the stored payloads for the tenant-scoped run. `POST /comparison-runs/final` creates a `final` run, stores `response_payload.snapshot`, enforces all submissions `ready` + readiness pass, and requires two ready quotes when `vendors_count >= 2` on the RFQ (from loaded attributes; otherwise at least one). `GET /comparison-runs` supports `?rfq_id=`. `AwardController` now validates award vendors against tenant-scoped RFQ submissions, persists per-vendor debrief rows in `Debrief`, serializes the frozen comparison vendor snapshot for award display, and keeps signoff idempotent in addition to creating/listing live award records. `POST /approvals/{id}/approve` gates on final run + snapshot + live readiness. `DecisionTrailRecorder::recordSnapshotFrozen` writes `comparison_snapshot_frozen` entries and now also records award debrief events. `DecisionTrailController` index/show read from `decision_trail_entries`. `PATCH /comparison-runs/{id}/scoring-model`, `POST /comparison-runs/{id}/lock`, and `POST /comparison-runs/{id}/unlock` now return explicit `422` beta-deferred responses (`COMPARISON_CONTROL_DEFERRED`) instead of mutating fake state. Tests: `ComparisonRunWorkflowTest`, `ComparisonSnapshotWorkflowTest`.
