+++ apps/atomy-q/API/app/Adapters/QuoteIngestion/EloquentNormalizationSourceLineRepository.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace App\Adapters\QuoteIngestion;
 
 use App\Models\NormalizationSourceLine;
-use Nexus\QuoteIngestion\Contracts\NormalizationSourceLineReadInterface;
-use Nexus\QuoteIngestion\Contracts\NormalizationSourceLineQueryInterface;
-use Nexus\QuoteIngestion\Contracts\NormalizationSourceLinePersistInterface;
+use App\Services\QuoteIntake\Contracts\NormalizationSourceLineReadInterface;
+use App\Services\QuoteIntake\Contracts\NormalizationSourceLineQueryInterface;
+use App\Services\QuoteIntake\Contracts\NormalizationSourceLinePersistInterface;
 
 final class EloquentNormalizationSourceLineRepository implements NormalizationSourceLineQueryInterface, NormalizationSourceLinePersistInterface
 {
