+++ apps/atomy-q/API/app/Adapters/QuoteIngestion/EloquentQuoteSubmissionPersist.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace App\Adapters\QuoteIngestion;
 
 use App\Models\QuoteSubmission;
-use Nexus\QuoteIngestion\Contracts\QuoteSubmissionInterface;
-use Nexus\QuoteIngestion\Contracts\QuoteSubmissionPersistInterface;
+use App\Services\QuoteIntake\Contracts\QuoteSubmissionInterface;
+use App\Services\QuoteIntake\Contracts\QuoteSubmissionPersistInterface;
 
 final readonly class EloquentQuoteSubmissionPersist implements QuoteSubmissionPersistInterface
 {
