+++ apps/atomy-q/API/app/Providers/AppServiceProvider.php
@@ -70,6 +70,11 @@ use App\Services\ProjectManagementOperations\AtomyLaborHealthService;
 use App\Services\ProjectManagementOperations\AtomyMilestoneBillingService;
 use App\Services\ProjectManagementOperations\AtomyProjectTaskIdsQuery;
 use App\Services\ProjectManagementOperations\AtomyTimelineDriftService;
+use App\Services\QuoteIntake\Contracts\NormalizationSourceLinePersistInterface;
+use App\Services\QuoteIntake\Contracts\NormalizationSourceLineQueryInterface;
+use App\Services\QuoteIntake\Contracts\QuoteSubmissionPersistInterface;
+use App\Services\QuoteIntake\Contracts\QuoteSubmissionQueryInterface;
+use App\Services\QuoteIntake\QuoteIngestionOrchestrator;
 use App\Services\SourcingOperations\AtomyRfqInvitationPersist;
 use App\Services\SourcingOperations\AtomyRfqInvitationQuery;
 use App\Services\SourcingOperations\AtomyRfqInvitationReminder;
@@ -156,11 +161,6 @@ use Nexus\TenantOperations\Contracts\TenantCreatorAdapterInterface;
 use Nexus\TenantOperations\Coordinators\TenantCompanyOnboardingCoordinator;
 use Nexus\TenantOperations\Services\TenantCompanyOnboardingService;
 use Nexus\Laravel\Idempotency\Contracts\ReplayResponseFactoryInterface;
-use Nexus\QuoteIngestion\QuoteIngestionOrchestrator;
-use Nexus\QuoteIngestion\Contracts\QuoteSubmissionQueryInterface;
-use Nexus\QuoteIngestion\Contracts\QuoteSubmissionPersistInterface;
-use Nexus\QuoteIngestion\Contracts\NormalizationSourceLineQueryInterface;
-use Nexus\QuoteIngestion\Contracts\NormalizationSourceLinePersistInterface;
 use Nexus\QuotationIntelligence\Contracts\QuotationIntelligenceCoordinatorInterface;
 use Nexus\QuotationIntelligence\Contracts\OrchestratorDocumentRepositoryInterface;
 use Nexus\QuotationIntelligence\Contracts\OrchestratorTenantRepositoryInterface;
