+++ apps/atomy-q/API/app/Services/QuoteIntake/Contracts/NormalizationSourceLineQueryInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Services\QuoteIntake\Contracts;
+
+interface NormalizationSourceLineQueryInterface
+{
+    public function findExisting(string $tenantId, string $quoteSubmissionId, string $rfqLineItemId): ?NormalizationSourceLineReadInterface;
+}
