+++ apps/atomy-q/API/app/Services/QuoteIntake/Contracts/QuoteSubmissionInterface.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Services\QuoteIntake\Contracts;
+
+interface QuoteSubmissionInterface
+{
+    public function getId(): string;
+
+    public function getTenantId(): string;
+
+    public function getRfqId(): string;
+
+    public function getVendorName(): string;
+}
