+++ apps/atomy-q/API/app/Services/QuoteIntake/Contracts/QuoteSubmissionQueryInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Nexus\QuoteIngestion\Contracts;
+namespace App\Services\QuoteIntake\Contracts;
 
 interface QuoteSubmissionQueryInterface
 {
