+++ apps/atomy-q/API/tests/Feature/QuoteIngestionIntelligenceTest.php
@@ -15,7 +15,7 @@ use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Str;
-use Nexus\QuoteIngestion\QuoteIngestionOrchestrator;
+use App\Services\QuoteIntake\QuoteIngestionOrchestrator;
 use Nexus\Tenant\Contracts\TenantContextInterface;
 use Tests\Feature\Api\ApiTestCase;
