--- orchestrators/QuoteIngestion/IMPLEMENTATION_SUMMARY.md
@@ -1,39 +1,0 @@
-# Implementation Summary: QuoteIngestion
-
-**Orchestrator:** `Nexus\QuoteIngestion`  
-**Status:** Active hardening  
-**Last Updated:** 2026-04-12
-
-## Scope of this hardening pass
-
-1. Removed Layer 2 framework coupling from this package:
-   - Dropped `laravel/framework` from `composer.json`.
-   - Removed `QuoteIngestionServiceProvider` from Layer 2 (service registration belongs in Layer 3 adapters).
-2. Hardened `QuoteIngestionOrchestrator` handling for malformed intelligence payloads:
-   - Safely handles non-array line entries.
-   - Skips lines with invalid or empty `rfq_line_id`.
-   - Uses persisted line count (not raw payload count) when marking completion.
-   - Computes average confidence from finite numeric values only.
-   - Normalizes scalar checks for source description, uom, taxonomy code, and mapping version.
-   - Sanitizes failure messages before persistence to avoid leaking internal error details.
-   - Exposes the generic failure message as a shared constant reused by the retry-exhaustion job path.
-3. Added unit tests for core orchestration behavior and failure paths.
-4. 2026-04-30 alpha readiness update:
-   - Normalizes provider-style fractional confidence values (`0..1`, inclusive) to the package readiness scale (`0..100`) before decision-trail and completion status decisions.
-   - Keeps deterministic confidence values already emitted on the `0..100` scale unchanged, and ignores values outside that scale.
-
-## Verification coverage added
-
-- Missing submission logs and exits safely.
-- Success flow:
-  - malformed lines are skipped,
-  - valid lines are persisted,
-  - decision trail writes only for qualified confidence,
-  - completion status and persisted line count are correct.
-- Coordinator failure marks submission as failed and clears tenant context.
-- Confidence averaging falls back to `0.0` when no finite numeric confidence is available.
-- Confidence scale detection:
-  - `0.0 -> 0.0`, `0.5 -> 50.0`, `0.95 -> 95.0`, `1.0 -> 100.0`.
-  - `1.5 -> 1.5`, `50 -> 50.0`; values greater than `1.0` and at most `100.0` are already percentage-scale.
-  - `-0.1` and `150` are outside the accepted readiness scale and are ignored for averaging.
-  - `ready` requires average normalized confidence `>= 80.0`; lower or missing valid confidence remains `needs_review`.
