--- orchestrators/QuoteIngestion/composer.json
@@ -1,42 +1,0 @@
-{
-    "name": "nexus/quote-ingestion",
-    "description": "Quote ingestion orchestration - extraction, normalization, persistence",
-    "type": "library",
-    "repositories": [
-        {
-            "type": "path",
-            "url": "../../packages/*",
-            "options": {
-                "symlink": true
-            }
-        },
-        {
-            "type": "path",
-            "url": "../../orchestrators/*",
-            "options": {
-                "symlink": true
-            }
-        }
-    ],
-    "require": {
-        "php": "^8.3",
-        "nexus/quotation-intelligence": "*@dev",
-        "azaharizaman/nexus-tenant": "*@dev",
-        "psr/log": "^3.0"
-    },
-    "require-dev": {
-        "phpunit/phpunit": "^10.0"
-    },
-    "autoload": {
-        "psr-4": {
-            "Nexus\\QuoteIngestion\\": "src/"
-        }
-    },
-    "autoload-dev": {
-        "psr-4": {
-            "Nexus\\QuoteIngestion\\Tests\\": "tests/"
-        }
-    },
-    "minimum-stability": "dev",
-    "prefer-stable": true
-}
