--- orchestrators/QuoteIngestion/src/Contracts/NormalizationSourceLinePersistInterface.php
@@ -1,15 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Nexus\QuoteIngestion\Contracts;
-
-interface NormalizationSourceLinePersistInterface
-{
-    public function upsert(
-        string $tenantId,
-        string $quoteSubmissionId,
-        string $rfqLineItemId,
-        array $data
-    ): void;
-}
