--- orchestrators/QuoteIngestion/src/Contracts/NormalizationSourceLineQueryInterface.php
@@ -1,14 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Nexus\QuoteIngestion\Contracts;
-
-interface NormalizationSourceLineQueryInterface
-{
-    public function findExisting(
-        string $tenantId,
-        string $quoteSubmissionId,
-        string $rfqLineItemId
-    ): ?NormalizationSourceLineReadInterface;
-}
